Use Text as the example backend in docstring
Fix typos in backend documentation

re parallelize tests

move deps back where they belong

fix parallel tests

typo

upgrade to pymc3 Beta!

add link for NUTS

commas

set transparency for histograms

make transforms classes

typo

add basic transform tests

actually basic test transforms

categorical foolings

improvements for simplex, but still broken

test and fix simplex jacobian

fix logtransform

test all jacobians

check transforms are going the right way

fix categorical

rename transforms

test and fix sum_to_1

fix unit continuous

try docker builds

remove sudo

Revert "remove sudo"

This reverts commit f2ce8353ed12c607e6bbd67615145d7711ab3bb3.

Revert "try docker builds"

This reverts commit b4d43ad823d900f99654679e0001f0f9e9e2b484.

added some docstrings

don't import nonexistant things

Combine chains by default

Make all methods for accessing variable values concatenate chains by
default, because this is likely to be the desired output most of the
time.

See #758 and #759.

Simplify single-trace __getitem__ method

The get_values fallback in the single-trace __getitem__ method was
present mostly as leftover code from a previous design.  MultiTrace is
the only user-facing trace object, so there is not much benefit of
letting __getitem__ fall back to get_values for convenience.

Allow trace variable values to be sliced

Extend indexing of trace objects to support an additional slice object
that specifies the burn and thin arguments of get_values:

    >>> trace[x, start::step]

Above differs from

    >>> trace[x][start::step]

because the second form operates on an array that is the combination
of all chains.

BaseTrace: Remove chain keyword argument to point

This is stale code.  The current BaseTrace only deals with a for a
chain argument in the point method.  The point method signatures in
the children classes are already correct.

MultiTrace: Rework docstring

* Add more information about getting variable values.

* Remove parameter information, which isn't relevant because users get
  back an initialized instance from sampling.

backends/base.py: Fix docstring typo

Differentiate names for single-chain traces

Previously, the variable name "trace" was confusingly used to refer to
both BaseTrace instances and MultiTrace instances, making readers
infer which type it was from context.

Change code to use the following conventions:

* Use "strace" for variable names of BaseTrace instances to indicate a
  single-chain trace (where "single-chain" trace refers to an object
  derived from BaseTrace, but not a MultiTrace object with only one
  chain).

* MultiTrace instances are called either "mtrace" or "trace".

All changes in this commit are purely variable renames and are not
user-facing.

Added alpha as an argument to traceplot

NDArray: Rename variable

Rename variable to make it clear that it refers to values for a single
model variable rather than a trace object.

https://github.com/pymc-devs/pymc3/pull/771#discussion_r32848583

Remove gridspec import check

Commit 3c3273da03 added a check on the gridspec import to support
older versions of matplotlib.  However, the minimum version that is
now specified (1.2.1) has gridspec, so this check is no longer needed.

combine leapfrog and energy computations

combine metropolis density comparison

cleanup things

revert some unintentional changes

avoid infinite recursion

actually fix arraystep

work with tensors too

shapes have to be ints

fix lkjcorr distribution

fastarraystep->arraystepshared

DOC Remove comment about TransformedVar in stoch vol example.

Removed TransformedVar from profiling example

Added standard normal cdf

Added ExGaussian distribution and (rubbish) tests

Added 2-parameter Inverse Gaussian distribution

Added 3-parameter shifted inverse gaussian.

Fix for phi parametrization in the inverse gaussian

Added newline at end of file.

Improved handling of alternative parametrization for inverse gaussian.

Added mean attribute to inverse gaussian.

Added switch to ExGaussian logp and missing self to method random.

Moved inverse gaussian stuff (and tests) to Wald.

Added condition to bound function in Wald logp

Fixed Wald logp

Fixed typos and removed whitespace

Remove unnecssary size checks from Wald/exGaussian random

Added disc_vars property to complement cont_vars. Might be useful for assigning discretes to a Metropolis sampler, for example.

MultiTrace: Check output before returning slice

Out-of-memory backends give a warning when the user tries to slice them,
so the result may be a list of Nones.

re: #790

Support slicing in SQlite and Text backends

Return a NDArray slice instead of warning when an out-of-memory backend
is sliced.

Fixes #790

ENH Add example of a Gaussian Mixture Model.

Add author

fix inappropriate summing

fix test for elemwise transform jacobian dets

remove debug statement

+++ pymc3/examples/LKJ_correlation.py
@@ -27,7 +27,7 @@ dataset = multivariate_normal(mu, cov_matrix, size=n_obs)
 
 # In order to convert the upper triangular correlation values to a complete
 # correlation matrix, we need to construct an index matrix:
-n_elem = n_var * (n_var - 1) / 2
+n_elem = int(n_var * (n_var - 1) / 2)
 tri_index = np.zeros([n_var, n_var], dtype=int)
 tri_index[np.triu_indices(n_var, k=1)] = np.arange(n_elem)
 tri_index[np.triu_indices(n_var, k=1)[::-1]] = np.arange(n_elem)
